python/requirements.txt
matplotlib==3.8.2
pyserial==3.5
